+++ AgentConnector/Editor/HeraAgentAssetConfigWindow.cs
@@ -155,7 +155,6 @@ namespace HeraAgent.Editor
 
             BuildToolbar();
             BuildUltraHeraSection();
-            BuildUiSystemSection();
             BuildGameFeelModeSection();
             BuildGameFeelUiModeSection();
             BuildUiSlopModeSection();